Heavy refactor of shadow/box-shadow logic in SleekElement, now creates 4 much smaller bitmaps to draw the edges of the view to scale better with large views and multiple shadow views, this also enables size animation that is fast enough not to affect frame rate.

+++ sleek_module/src/main/java/com/ztory/lib/sleek/animation/SAVtransXYWH.java
@@ -71,8 +71,8 @@ public class SAVtransXYWH implements ISleekAnimView {
         if (progressTs < duration) {
             percent = (float) progressTs / (float) duration;
             percent = percent * percent;
-            currTransX = startX + (amountTransX * percent);
-            currTransY = startY + (amountTransY * percent);
+            currTransX = startX + (int) (amountTransX * percent);
+            currTransY = startY + (int) (amountTransY * percent);
             currTransW = startW + (int) (amountTransW * percent);
             currTransH = startH + (int) (amountTransH * percent);
         }
